Updated Announcement Box
Announcement system now displays the announcements in manner decided. I
added sql so that it displays only announcements relevant to the user
(student in this case). Still need to do it specific professor and ta.
I need to figure out a way to add/post to database and a way to reduce
posts shown to a set number. Also need to figure out a way to count
correctly in id in db. Atm, the system works perfectly with the info in
the db.